epictrack-api/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
attrs==24.2.0
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dpath==2.2.0
ecdsa==0.19.0
et-xmlfile==1.1.0
Flask==2.2.5
Flask-Caching==2.3.0
flask-jwt-oidc==0.3.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
fonttools==4.54.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
Inflector==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.17.0
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.4
natsort==8.4.0
numpy==1.26.4
openpyxl==3.0.10
packaging==24.1
pandas==2.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
referencing==0.35.1
reportlab==3.6.12
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.3.6
XlsxWriter==3.0.6
